lib: move ssd1320 to separate folder
